refactor: float type change to bigdecimal

+++ init/create-database.sql
@@ -4,9 +4,9 @@ USE conversiondb;
 CREATE TABLE conversion_coin (
     id BINARY(16) NOT NULL,
     base VARCHAR(255) NOT NULL,
-    amount FLOAT(23) NOT NULL,
+    amount DECIMAL(23, 5) NOT NULL,
     currency VARCHAR(255) NOT NULL,
-    converted FLOAT NOT NULL,
+    converted DECIMAL(23, 5) NOT NULL,
     date DATE NOT NULL,
     PRIMARY KEY (id)
 );
@@ -17,9 +17,9 @@ USE conversiondb_test;
 CREATE TABLE conversion_coin (
     id BINARY(16) NOT NULL,
     base VARCHAR(255) NOT NULL,
-    amount FLOAT(23) NOT NULL,
+    amount DECIMAL(23, 5) NOT NULL,
     currency VARCHAR(255) NOT NULL,
-    converted FLOAT NOT NULL,
+    converted DECIMAL(23, 5) NOT NULL,
     date DATE NOT NULL,
     PRIMARY KEY (id)
 );
