+++ src/main/java/br/inatel/conversionmanager/model/dto/ExchangeRateResponse.java
@@ -2,13 +2,14 @@ package br.inatel.conversionmanager.model.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Builder
 public record ExchangeRateResponse(long timestamp,
-                           String base,
-                           boolean success,
-                           Map<String, Float> rates,
-                           String date,
-                           boolean historical) {
+                                   String base,
+                                   boolean success,
+                                   Map<String, BigDecimal> rates,
+                                   String date,
+                                   boolean historical) {
 }
